Additional fix for #4 - static cast octree leafs count to uint_fast32_t

+++ Source/Nav3D/Private/Nav3DVolume.cpp
@@ -465,7 +465,7 @@ void ANav3DVolume::GetAdjacentLeafs(const FNav3DOctreeEdge& Edge, TArray<FNav3DO
 					continue;
 				}
 				
-				if (AdjacentNode.FirstChild.NodeIndex < Octree.Leafs.Num()) {
+				if (AdjacentNode.FirstChild.NodeIndex < static_cast<uint_fast32_t>(Octree.Leafs.Num())) {
 					const FNav3DOctreeLeaf& Leaf = Octree.Leafs[AdjacentNode.FirstChild.NodeIndex];
 
 					if (!Leaf.IsOccluded()) {
